requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ML models
xgboost>=2.0.0
lightgbm>=4.0.0

# Data fetching
yfinance>=0.2.28
requests>=2.31.0
finnhub-python>=2.4.18

# Sentiment analysis
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Technical indicators
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional (for Colab)
ipywidgets>=8.0.0


